desktop: Add missing x-ms-wma MIME type
Currently, the audio/x-ms-wma MIME type is not advertised as supported.
This results in WMA audio files being opened by Showtime by default,
since it advertises support for video/x-ms-asf, which is a catch-all for
Windows Media files containing audio and video under the "Advanced
System Format" family name.

Decibels can open WMA files, so advertise support for it in the desktop
file, so that those files are opened by default by the audio player,
instead of the video player.